+++ api/cf-node-client/src/lib.rs
@@ -28,3 +28,28 @@ pub enum WaitForResult {
 	TransactionHash(H256),
 	Details(ExtrinsicDetails),
 }
+
+#[derive(Debug, Serialize, Deserialize, Clone)]
+#[serde(rename_all = "snake_case")]
+pub enum ApiWaitForResult<T> {
+	TxHash(H256),
+	TxDetails { tx_hash: H256, response: T },
+}
+
+impl<T> ApiWaitForResult<T> {
+	pub fn map_details<R>(self, f: impl FnOnce(T) -> R) -> ApiWaitForResult<R> {
+		match self {
+			ApiWaitForResult::TxHash(hash) => ApiWaitForResult::TxHash(hash),
+			ApiWaitForResult::TxDetails { response, tx_hash } =>
+				ApiWaitForResult::TxDetails { tx_hash, response: f(response) },
+		}
+	}
+
+	#[track_caller]
+	pub fn unwrap_details(self) -> T {
+		match self {
+			ApiWaitForResult::TxHash(_) => panic!("unwrap_details called on TransactionHash"),
+			ApiWaitForResult::TxDetails { response, .. } => response,
+		}
+	}
+}
